requirements.txt
numpy==1.26.4
beautifulsoup4
fastapi
faiss-cpu
pdfplumber
scikit-learn
sentence-transformers
uvicorn
llama-cpp-python
python-multipart